+++ tests/test_registry.py
@@ -2,8 +2,6 @@
 Tests for the SerializableRegistry class.
 """
 
-import pytest
-
 from serilux import Serializable, SerializableRegistry, register_serializable
 
 
@@ -45,15 +43,17 @@ class TestSerializableRegistry:
 
         assert SerializableRegistry.get_class("MyClass") == MyClass
 
-    def test_register_serializable_validation(self, clear_registry):
-        """Test that @register_serializable validates __init__ signature."""
-        with pytest.raises(TypeError):
+    def test_register_serializable_with_required_args(self, clear_registry):
+        """Test that @register_serializable now allows required __init__ parameters."""
 
-            @register_serializable
-            class InvalidClass(Serializable):
-                def __init__(self, required_param):
-                    super().__init__()
-                    self.add_serializable_fields([])
+        @register_serializable
+        class RequiredArgsClass(Serializable):
+            def __init__(self, required_param):
+                super().__init__()
+                self.required_param = required_param
+                self.add_serializable_fields(["required_param"])
+
+        assert SerializableRegistry.get_class("RequiredArgsClass") == RequiredArgsClass
 
     def test_register_serializable_with_defaults(self, clear_registry):
         """Test that @register_serializable allows default parameters."""